Lesson 6

+++ app/controllers/answers_controller.rb
@@ -8,21 +8,19 @@ class AnswersController < ApplicationController
   def create
     @exposed_answer = question.answers.new(answer_params)
     answer.author = current_user
+    answer.save
+  end
 
-    if answer.save
-      redirect_to question, notice: "Your answer was successfully created."
-    else
-      render "questions/show"
-    end
+  def update
+    answer.update(answer_params) if current_user.author_of?(answer)
   end
 
   def destroy
-    if current_user.author_of?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: "Answer was successfully deleted"
-    else
-      redirect_to answer.question, notice: "Can't delete the answer"
-    end
+    answer.destroy if current_user.author_of?(answer)
+  end
+
+  def best
+    answer.set_best if current_user.author_of?(answer.question)
   end
 
   private
